requirements.txt
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
decorator==4.4.2
Flask==1.1.2
Flask-Mail==0.9.1
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
moviepy==1.0.3
netifaces==0.10.6
numpy==1.18.4
passlib==1.7.4
Pillow==7.1.2
proglog==0.1.9
PyJWT==2.1.0
requests==2.23.0
six==1.14.0
tqdm==4.46.0
urllib3==1.25.9
Werkzeug==1.0.1
